# app/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
...
        "title": "pint-server",
        "contact": {},
        "license": {
            "name": "MIT"
        }
    },
    "paths": {
...
        "/v3/envelopes/{envelopeReference}/finish-transfer": {
            "put": {
                "description": "Use this endpoint after you have finished transferring all additional documents.\n\nPrior to accepting envelope transfer, the receiving platform ensures that all supporting documents listed in the envelope manifest have been successfully transferred.\n\n**Success Responses**\n\n`200 OK` - Transfer accepted (RECE)\n\nAll additional documents have been received and the envelope transfer has been accepted on the platform.\n\nThe payload of the signed response contains the last transfer chain entry checksum and a list of all received additional document checksums.\n(See the `default` response for details of the response payload.)\n\n**Retry handling:**\nWhen the sender retries a transfer that has already been accepted, the receiver will return a signed response and\nthe payload will contain a structure identical to the original response, but with a response code of DUPE\nand an extra field: duplicateOfAcceptedEnvelopeTransferChainEntrySignedContent.\n\n**Notes**\n\n**Error Responses (signed)**\n\n`409 Conflict` - Missing documents (MDOC)\n- One or more additional documents are still missing (see missingAdditionalDocumentChecksums in response payload).\n\n`422 Unprocessable Entity` - Envelope rejected (BSIG/BENV) (signed response)\n- BSIG: Signature validation failed.\n- BENV: Envelope validation failed.\n\nSee the `default` response for details of the response payload.\n\n**Error Responses (unsigned)**\n\n`400 Bad Request` - Malformed request - returned as an unsigned error response.\n`500 Internal Server Error` - Internal error - returned as an unsigned error response.\n\n**Note:** Unsigned responses cannot be verified as originating from the receiving platform.\nDo not assume an unsigned error response means the transfer was rejected.\nOnly determine transfer acceptance/rejection from signed responses.\n",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "PINT"
                ],
                "summary": "Finish envelope transfer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Envelope reference (UUID)",
                        "name": "envelopeReference",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signed response - Transfer accepted (RECE/DUPE)",
                        "schema": {
                            "$ref": "#/definitions/pint.SignedEnvelopeTransferFinishedResponse"
                        }
                    },
                    "409": {
                        "description": "Signed response - Missing documents (MDOC) or disputed (DISE)",
                        "schema": {
                            "$ref": "#/definitions/pint.SignedEnvelopeTransferFinishedResponse"
                        }
                    },
                    "422": {
                        "description": "Signed response - Envelope rejected (BSIG/BENV)",
                        "schema": {
                            "$ref": "#/definitions/pint.SignedEnvelopeTransferFinishedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal error processing request",
                        "schema": {
                            "$ref": "#/definitions/pint.ErrorResponse"
                        }
                    },
                    "default": {
                        "description": "documentation only (not returned directly) - decoded payload of the signed response",
                        "schema": {
                            "$ref": "#/definitions/pint.EnvelopeTransferFinishedResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pint.DetailedError": {
            "type": "object",
            "properties": {
                "errorCode": {
                    "description": "error code used on the platform: 7000-7999 for technical errors, 8000-8999 for functional errors",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pint.ErrorCode"
                        }
                    ]
                },
                "errorCodeMessage": {
                    "type": "string"
                },
                "errorCodeText": {
                    "type": "string"
                },
                "jsonPath": {
                    "type": "string"
                },
                "property": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "pint.EnvelopeTransferFinishedResponse": {
            "type": "object",
            "properties": {
                "duplicateOfAcceptedEnvelopeTransferChainEntrySignedContent": {
                    "description": "DuplicateOfAcceptedEnvelopeTransferChainEntrySignedContent is the last transfer chain entry\nfrom the previously accepted envelope transfer.\nOnly included when ResponseCode is DUPE.\nThis is a JWS compact serialization token.",
                    "type": "string",
                    "example": "eyJhbGciOiJFZERTQSIsImtpZCI6IjQ0MzkzNDQ5MzQ0OTM0NDkzNDQ5MzQ0OTM0NDkzNDQ5MzQ0OTM0NDkzNDQ5MzQ0OTM0NDkzNDQ5MzQ0OTM0NDkzNDQ5Mz"
                },
                "lastEnvelopeTransferChainEntrySignedContentChecksum": {
                    "description": "LastEnvelopeTransferChainEntrySignedContentChecksum is the SHA-256 checksum of the last\ntransfer chain entry received.",
                    "type": "string",
                    "example": "20a0257b313ae08417e07f6555c4ec829a512c083f3ead16b41158018a22abe9"
                },
                "missingAdditionalDocumentChecksums": {
                    "description": "MissingAdditionalDocumentChecksums lists the checksums of additional documents that\nhave not been received by the receiving platform.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "583c29ab3e47f2d80899993200d3fbadb9f8a367f3a39f715935c46d7a283006"
                    ]
                },
                "reason": {
                    "description": "Reason is a free text comment clarifying the result or suggesting follow-up actions.\nOmitted when ResponseCode is RECE (no additional information needed).",
                    "type": "string",
                    "example": "jws.Verify(): invalid key type"
                },
                "receivedAdditionalDocumentChecksums": {
                    "description": "ReceivedAdditionalDocumentChecksums confirms all additional documents received during\nthe envelope transfer.\nIncluded with RECE or DUPE ResponseCode to provide a signed receipt.\nThis includes all additional documents (including ones the receiver already had).",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "123329ab3e47f2d80899993200d3fbadb9f8a367f3a39f715935c46d7a283006"
                    ]
                },
                "responseCode": {
                    "description": "ResponseCode indicates the result of the envelope transfer.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pint.ResponseCode"
                        }
                    ],
                    "example": "BSIG"
                }
            }
        },
...
        "pint.ErrorCode": {
            "type": "integer",
            "enum": [
                7001,
                7002,
                7003,
                7004,
                7005,
                7006,
                7007,
                7008,
                7009,
                7010,
                7011,
                8001,
                8002
            ],
            "x-enum-comments": {
                "ErrCodeInsufficientTrust": "Trust level below minimum required",
                "ErrCodeMalformedRequest": "JSON parsing or encoding failed"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "JSON parsing or encoding failed",
                "",
                "",
                "",
                "",
                "",
                "",
                "Trust level below minimum required"
            ],
            "x-enum-varnames": [
                "ErrCodeBadSignature",
                "ErrCodeBadCertificate",
                "ErrCodeBadChecksum",
                "ErrCodeInvalidEnvelope",
                "ErrCodeInternalError",
                "ErrCodeMalformedRequest",
                "ErrCodeKeyError",
                "ErrCodeRegistryError",
                "ErrCodeRateLimitExceeded",
                "ErrCodeRequestTooLarge",
                "ErrCodeNotFound",
                "ErrCodeUnknownParty",
                "ErrCodeInsufficientTrust"
            ]
        },
        "pint.ErrorResponse": {
            "type": "object",
            "properties": {
                "errorDateTime": {
                    "description": "The DateTime corresponding to the error occurring",
                    "type": "string"
                },
                "errors": {
                    "description": "An array of errors providing more detail about the root cause",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pint.DetailedError"
                    }
                },
                "httpMethod": {
                    "description": "The HTTP method used to make the request e.g. GET, POST, etc",
                    "type": "string"
                },
                "providerCorrelationReference": {
                    "description": "A unique identifier to the HTTP request within the scope of the API provider",
                    "type": "string"
                },
                "requestUri": {
                    "description": "The URI that was requested",
                    "type": "string"
                },
                "statusCode": {
                    "description": "The HTTP status code returned",
                    "type": "integer"
                },
                "statusCodeMessage": {
                    "description": "A long description corresponding to the HTTP status code with additional information",
                    "type": "string"
                },
                "statusCodeText": {
                    "description": "A standard short description corresponding to the HTTP status code",
                    "type": "string"
                }
            }
        },
...
        "pint.ResponseCode": {
            "type": "string",
            "enum": [
                "RECE",
                "DUPE",
                "BSIG",
                "BENV",
                "INCD",
                "MDOC"
            ],
            "x-enum-varnames": [
                "ResponseCodeRECE",
                "ResponseCodeDUPE",
                "ResponseCodeBSIG",
                "ResponseCodeBENV",
                "ResponseCodeINCD",
                "ResponseCodeMDOC"
            ]
        },
        "pint.SignedEnvelopeTransferFinishedResponse": {
            "type": "object",
            "properties": {
                "envelopeTransferFinishedResponseSignedContent": {
                    "description": "EnvelopeTransferFinishedResponseSignedContent is a JWS-signed response returned when\nan envelope transfer is accepted or rejected.",
                    "type": "string",
                    "example": "eyJhbGciOiJFZERTQSIsImtpZCI6IjQ0MzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkMzlkM"
                }
            }
        },
...
    },
...
}